+++ crates/common/fmt/src/ui.rs
@@ -7,13 +7,12 @@ use alloy_network::{
     AnyHeader, AnyReceiptEnvelope, AnyRpcBlock, AnyRpcTransaction, AnyTransactionReceipt,
     AnyTxEnvelope, ReceiptResponse,
 };
-use alloy_primitives::{
-    Address, Bloom, Bytes, FixedBytes, I256, Log as PrimitiveLog, U8, U64, U256, Uint, hex,
-};
+use alloy_primitives::{Address, Bloom, Bytes, FixedBytes, I256, U8, U64, U256, Uint, hex};
 use alloy_rpc_types::{
     AccessListItem, Block, BlockTransactions, Header, Log, Transaction, TransactionReceipt,
 };
 use alloy_serde::{OtherFields, WithOtherFields};
+use foundry_primitives::{FoundryReceiptEnvelope, FoundryTxReceipt};
 use revm::context_interface::transaction::SignedAuthorization;
 use serde::Deserialize;
 
@@ -258,20 +257,6 @@ transactionIndex: {}",
             self.topics().pretty(),
             self.transaction_hash.pretty(),
             self.transaction_index.pretty(),
-        )
-    }
-}
-
-impl UIfmt for PrimitiveLog {
-    fn pretty(&self) -> String {
-        format!(
-            "
-address: {}
-data: {}
-topics: {}",
-            self.address.pretty(),
-            self.data.data.pretty(),
-            self.topics().iter().map(|t| t.pretty()).collect::<Vec<_>>().join(", ")
         )
     }
 }
@@ -837,6 +822,94 @@ impl UIfmt for SignedAuthorization {
                 "{{recoveredAuthority: <error: {e}>, signedAuthority: {signed_authorization}}}",
             ),
         }
+    }
+}
+
+impl<T> UIfmt for FoundryReceiptEnvelope<T>
+where
+    T: UIfmt + Clone + core::fmt::Debug + PartialEq + Eq,
+{
+    fn pretty(&self) -> String {
+        let receipt = self.as_receipt();
+        let deposit_info = match self {
+            Self::Deposit(d) => {
+                format!(
+                    "
+depositNonce         {}
+depositReceiptVersion {}",
+                    d.receipt.deposit_nonce.pretty(),
+                    d.receipt.deposit_receipt_version.pretty()
+                )
+            }
+            _ => String::new(),
+        };
+
+        format!(
+            "
+status               {}
+cumulativeGasUsed    {}
+logs                 {}
+logsBloom            {}
+type                 {}{}",
+            receipt.status.pretty(),
+            receipt.cumulative_gas_used.pretty(),
+            receipt.logs.pretty(),
+            self.logs_bloom().pretty(),
+            self.tx_type() as u8,
+            deposit_info
+        )
+    }
+}
+
+impl UIfmt for FoundryTxReceipt {
+    fn pretty(&self) -> String {
+        let receipt = &self.0.inner;
+        let other = &self.0.other;
+
+        let mut pretty = format!(
+            "
+blockHash            {}
+blockNumber          {}
+contractAddress      {}
+cumulativeGasUsed    {}
+effectiveGasPrice    {}
+from                 {}
+gasUsed              {}
+logs                 {}
+logsBloom            {}
+root                 {}
+status               {}
+transactionHash      {}
+transactionIndex     {}
+type                 {}
+blobGasPrice         {}
+blobGasUsed          {}",
+            receipt.block_hash.pretty(),
+            receipt.block_number.pretty(),
+            receipt.contract_address.pretty(),
+            receipt.inner.cumulative_gas_used().pretty(),
+            receipt.effective_gas_price.pretty(),
+            receipt.from.pretty(),
+            receipt.gas_used.pretty(),
+            serde_json::to_string(receipt.inner.logs()).unwrap(),
+            receipt.inner.logs_bloom().pretty(),
+            self.state_root().pretty(),
+            receipt.inner.status().pretty(),
+            receipt.transaction_hash.pretty(),
+            receipt.transaction_index.pretty(),
+            receipt.inner.tx_type() as u8,
+            receipt.blob_gas_price.pretty(),
+            receipt.blob_gas_used.pretty()
+        );
+
+        if let Some(to) = receipt.to {
+            pretty.push_str(&format!("\nto                   {}", to.pretty()));
+        }
+
+        // additional captured fields
+        pretty.push_str(&other.pretty());
+
+        pretty
     }
 }
 
@@ -1628,5 +1701,35 @@ validBefore          null
 "#
                 .trim()
         );
+    }
+
+    #[test]
+    fn test_foundry_tx_receipt_uifmt() {
+        use alloy_network::AnyTransactionReceipt;
+        use foundry_primitives::FoundryTxReceipt;
+
+        // Test UIfmt implementation for FoundryTxReceipt
+        let s = r#"{"type":"0x2","status":"0x1","cumulativeGasUsed":"0x5208","logs":[],"logsBloom":"0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000","transactionHash":"0x1234567890123456789012345678901234567890123456789012345678901234","transactionIndex":"0x0","blockHash":"0xabcdefabcdefabcdefabcdefabcdefabcdefabcdefabcdefabcdefabcdefabcd","blockNumber":"0x1","gasUsed":"0x5208","effectiveGasPrice":"0x3b9aca00","from":"0x1234567890123456789012345678901234567890","to":"0x0987654321098765432109876543210987654321","contractAddress":null}"#;
+        let any_receipt: AnyTransactionReceipt = serde_json::from_str(s).unwrap();
+        let foundry_receipt = FoundryTxReceipt::try_from(any_receipt).unwrap();
+
+        let pretty_output = foundry_receipt.pretty();
+
+        // Check that essential fields are present in the output
+        assert!(pretty_output.contains("blockHash"));
+        assert!(pretty_output.contains("blockNumber"));
+        assert!(pretty_output.contains("status"));
+        assert!(pretty_output.contains("gasUsed"));
+        assert!(pretty_output.contains("transactionHash"));
+        assert!(pretty_output.contains("type"));
+
+        // Verify the transaction hash appears in the output
+        assert!(
+            pretty_output
+                .contains("0x1234567890123456789012345678901234567890123456789012345678901234")
+        );
+
+        // Verify status is pretty printed correctly (boolean true for successful transaction)
+        assert!(pretty_output.contains("true"));
     }
 }
