+++ crates/primitives/Cargo.toml
@@ -29,4 +29,3 @@ serde_json.workspace = true
 serde = { version = "1.0", features = ["derive"] }
 derive_more.workspace = true
 tempo-primitives.workspace = true
-foundry-common-fmt.workspace = true
